feat: Add Limit and Quota
This adds two core structs, Limit and Quota. The rate limiter will be
given Limits that define how many requests to perform an action against
a resource can be made within a time period. The limiter will then
create and used Quotas to track the number of remaining requests that
can be made by a given requester.