feat(limiter): Add method to set HTTP rate limit policy header
This adds the concept of a rate limit policy. A limit policy is a
collection of Limits for the same combination of resource and action,
one Limit per IP address, auth token, and in total.

The limiter can now be used to set an HTTP header to provide the policy
for a given resource and action as a response header. This allows for
informing clients about the rate limit policy so they can adapt their
rate of requests prior to reaching a limit. By default the limiter will
use the RateLimit-Policy header. This can be changed using the
WithPolicyHeader option when creating the limiter.