feat(limiter): Add method to set HTTP rate limit usage header
The limiter can now be used to set an HTTP header to provide information
about the current quota usage for a given resource and action. This
allows for informing clients about the current quota usage so they can
adapt their rate of requests prior to reaching a limit. By default the
limiter will use the RateLimit header. This can be changed using the
WithUsageHeader option when creating the limiter.