refact: Use separate types to differentiate limited and unlimited
This splits the concept of an Unlimited limit, and a Limited limit into
two separate types rather than a boolean field on a Limit. It changes
Limit into an interface that both Unlimited and Limited implement.

This simplifies the validation logic of a limit, and will make it
clearer when a limit is meant to be Unlimited.